requirements.txt
streamlit
langchain
langchain-huggingface
langchain-community
pypdf2
faiss-cpu
sentence-transformers
python-dotenv